// packages/core/tests/fixtures/search/zenstack/schema.ts
//////////////////////////////////////////////////////////////////////////////////////////////
// DO NOT MODIFY THIS FILE                                                                  //
// This file is automatically generated by ZenStack CLI and should not be manually updated. //
//////////////////////////////////////////////////////////////////////////////////////////////

/* eslint-disable */

import { type SchemaDef, type AttributeApplication, type FieldDefault, ExpressionUtils } from "@zenstackhq/schema";
export class SchemaType implements SchemaDef {
    provider = {
        type: "postgresql"
    } as const;
    models = {
        Author: {
            name: "Author",
            fields: {
                id: {
                    name: "id",
                    type: "String",
                    id: true,
                    attributes: [{ name: "@id" }, { name: "@default", args: [{ name: "value", value: ExpressionUtils.call("cuid") }] }] as readonly AttributeApplication[],
                    default: ExpressionUtils.call("cuid") as FieldDefault
                },
                name: {
                    name: "name",
                    type: "String"
                },
                articles: {
                    name: "articles",
                    type: "Article",
                    array: true,
                    relation: { opposite: "author" }
                }
            },
            attributes: [
                { name: "@@allow", args: [{ name: "operation", value: ExpressionUtils.literal("all") }, { name: "condition", value: ExpressionUtils.literal(true) }] }
            ] as readonly AttributeApplication[],
            idFields: ["id"],
            uniqueFields: {
                id: { type: "String" }
            }
        },
        Article: {
            name: "Article",
            fields: {
                id: {
                    name: "id",
                    type: "String",
                    id: true,
                    attributes: [{ name: "@id" }, { name: "@default", args: [{ name: "value", value: ExpressionUtils.call("cuid") }] }] as readonly AttributeApplication[],
                    default: ExpressionUtils.call("cuid") as FieldDefault
                },
                title: {
                    name: "title",
                    type: "String",
                    fuzzy: true,
                    fullText: true,
                    attributes: [{ name: "@fuzzy" }, { name: "@fullText" }] as readonly AttributeApplication[]
                },
                body: {
                    name: "body",
                    type: "String",
                    fullText: true,
                    attributes: [{ name: "@fullText" }] as readonly AttributeApplication[]
                },
                author: {
                    name: "author",
                    type: "Author",
                    attributes: [{ name: "@relation", args: [{ name: "fields", value: ExpressionUtils.array("String", [ExpressionUtils.field("authorId")]) }, { name: "references", value: ExpressionUtils.array("String", [ExpressionUtils.field("id")]) }] }] as readonly AttributeApplication[],
                    relation: { opposite: "articles", fields: ["authorId"], references: ["id"] }
                },
                authorId: {
                    name: "authorId",
                    type: "String",
                    foreignKeyFor: [
                        "author"
                    ] as readonly string[]
                }
            },
            attributes: [
                { name: "@@allow", args: [{ name: "operation", value: ExpressionUtils.literal("all") }, { name: "condition", value: ExpressionUtils.literal(true) }] }
            ] as readonly AttributeApplication[],
            idFields: ["id"],
            uniqueFields: {
                id: { type: "String" }
            }
        }
    } as const;
    plugins = {};
}
export const schema = new SchemaType();